// src/interfaces/student.interface.ts
export interface StudentInt {
  first_name: string;
  last_name: string;
  email: string;
  password: string;
  gender: string;
  dob: string;
  nationality: string;
  highest_qualifications: string;
  phone: string;
  city: string;
  sponsor_name: string;
  sponsor_email: string;
  sponsor_phone: string;
}

export interface User {
  name: string;
  email: string;
  password: string;
}